// src/ImageHotspot/Client/src/types/imagehotspotvalue.d.ts
export class ImageHotspotValue {
  image?: string | null = null | undefined;
  width: number | null = null | undefined;
  height: number | null = null | undefined;
  percentX: number | null = null | undefined;
  percentY: number | null = null | undefined;
  left: number | null = null | undefined;
  top: number | null = null | undefined;
}